requirements.txt
customtkinter==5.2.2
numpy==1.26.4
pillow==10.2.0
sympy==1.12
tk==0.1.0
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
